requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sse-starlette>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Redis / Queue
redis>=5.0.0
celery[redis]>=5.3.0

# Object Storage
boto3>=1.34.0

# AI SDKs
google-genai>=1.0.0
openai>=1.12.0

# Content Extraction
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
youtube-transcript-api>=0.6.0
yt-dlp>=2024.1.0
playwright>=1.41.0

# Video / Image
moviepy>=1.0.3
Pillow>=10.2.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Resilience
tenacity>=8.2.0

# Observability
loguru>=0.7.0
prometheus-client>=0.20.0

# Config
pydantic-settings>=2.1.0
pydantic>=2.6.0

# NLP (Evidence Pack)
scikit-learn>=1.4.0

# Utils
python-multipart>=0.0.6
httpx>=0.27.0

# Test
pytest>=8.0.0
pytest-asyncio>=0.23.0
